pyproject.toml
[tool.poetry]
name = "Monarch-Technical-Documentation"
version = "0.1.1"
description = ""
authors = [
    "Monarch Initiative <info@monarchinitiative.org>",
    "Glass Elsarboukh <glass@tislab.org>",
    "Tim Putman <tim@tislab.org>",
    "Kevin Schaper <kevin@tislab.org>",
    "Corey Cox <corey@tislab.org>"
]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.1"
PyYAML = "^6.0"
PyGithub = "^1.55"
linkml = "^1.3.7"
linkml-runtime = "^1.1.24"
click = "8.0.4"
cruft = "^2.11.1"
mkdocs-material = "^8.2.8"
mkdocs-mermaid2-plugin = "^0.6.0"
j2cli = "^0.3.10"
loguru = "*"

[tool.poetry.extras]
j2cli = ["yaml"]

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"

[tool.poetry.scripts]

[tool.poetry.group.dev.dependencies]
linkml = "1.3.13"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
